+++ init.sql
@@ -42,7 +42,7 @@ CREATE TABLE round (
                        id BIGINT PRIMARY KEY AUTO_INCREMENT,
                        party_id BIGINT,
                        game_id BIGINT,
-                       answer INT,
+                       answer BIGINT,
                        created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
                         terminated_at DATETIME
 );
